+++ tests/integration/test_xtce_based_parsing/test_ctim_parsing.py
@@ -1,7 +1,7 @@
 """Test parsing of CTIM instrument data"""
 import pytest
 
-# Local
+from space_packet_parser import ccsds
 from space_packet_parser.xtce import definitions
 
 
@@ -16,10 +16,9 @@ def test_ctim_parsing(ctim_test_data_dir):
     print("Loading and parsing data")
     test_packet_file = ctim_test_data_dir / 'ccsds_2021_155_14_39_51'
     with open(test_packet_file, 'rb') as pkt_file:
-        pkt_gen = pkt_def.packet_generator(pkt_file,
-                                           root_container_name="CCSDSTelemetryPacket",
-                                           show_progress=True)
-        packets = list(pkt_gen)
+        ccsds_gen = ccsds.ccsds_generator(pkt_file, show_progress=True)
+        packets = [pkt_def.parse_bytes(binary_data, root_container_name="CCSDSTelemetryPacket")
+                   for binary_data in ccsds_gen]
 
     assert len(packets) == 1499
     assert packets[159]['PKT_APID'].raw_value == 34