+++ docs/source/changelog.md
@@ -11,6 +11,9 @@ Release notes for the `space_packet_parser` library
 - *BREAKING*: Reorganization of the project into different submodules for more explicit handling
   of imports. There is now an `space_packet_parser.xtce` module with xtce representations separated
   into modules underneath that.
+- *BREAKING*: Removed mid-level abstraction methods `packet_generator()` and `ccsds_packet_generator()`
+  from `XtcePacketDefinition`. Use low-level `parse_bytes()` with bytes generators directly, or high-level
+  `space_packet_parser.xarr.create_dataset()` for xarray integration.
 - Add support for creating a packet definition from Python objects and serializing it as XML.
 - BUGFIX: Fix kbps calculation in packet generator for showing progress.
 - Add support for string and float encoded enumerated lookup parameters.